Rename models to schemas, dm_models to models to reflect latest fastAPI docs. See tiangolo/full-stack-fastapi-postgresql#23
