+++ backend/app/app/api/api_v1/endpoints/farms/farms.py
@@ -8,7 +8,7 @@ from app import crud
 from app.api.utils.db import get_db
 from app.api.utils.farms import get_farm_client, ClientError, get_farms_url_or_list, get_farm_by_id
 from app.api.utils.security import get_farm_access
-from app.models.farm import Farm, FarmCreate, FarmUpdate
+from app.schemas.farm import Farm, FarmCreate, FarmUpdate
 from app.core.celery_app import celery_app
 
 router = APIRouter()
