+++ backend/app/app/api/api_v1/endpoints/farms/info.py
@@ -8,8 +8,8 @@ from app import crud
 from app.api.utils.db import get_db
 from app.api.utils.farms import get_farm_client, ClientError, get_farms_url_or_list
 from app.api.utils.security import get_farm_access
-from app.models.farm import Farm
-from app.models.farm_info import FarmInfo
+from app.schemas.farm import Farm
+from app.schemas.farm_info import FarmInfo
 
 router = APIRouter()
