+++ backend/app/app/db/base.py
@@ -1,6 +1,6 @@
-# Import all the models, so that Base has them before being
+# Import all the schemas, so that Base has them before being
 # imported by Alembic
 from app.db.base_class import Base  # noqa
-from app.db_models.user import User  # noqa
-from app.db_models.farm import Farm
-from app.db_models.farm_token import FarmToken
+from app.models.user import User  # noqa
+from app.models.farm import Farm
+from app.models.farm_token import FarmToken
