+++ backend/app/app/db/init_db.py
@@ -1,8 +1,8 @@
 from app import crud
 from app.core import config
-from app.models.user import UserCreate
+from app.schemas.user import UserCreate
 
-# make sure all SQL Alchemy models are imported before initializing DB
+# make sure all SQL Alchemy schemas are imported before initializing DB
 # otherwise, SQL Alchemy might fail to initialize properly relationships
 # for more details: https://github.com/tiangolo/full-stack-fastapi-postgresql/issues/28
 from app.db import base
