--- backend/app/app/models/msg.py
@@ -1,5 +1,0 @@
-from app.models.api_model import APIModel
-
-
-class Msg(APIModel):
-    msg: str
