+++ backend/app/app/schemas/msg.py
@@ -1,0 +1,5 @@
+from app.schemas.api_model import APIModel
+
+
+class Msg(APIModel):
+    msg: str
