+++ backend/app/app/schemas/user.py
@@ -1,0 +1,35 @@
+from typing import Optional
+
+from app.schemas.api_model import APIModel
+
+
+# Shared properties
+class UserBase(APIModel):
+    email: Optional[str] = None
+    is_active: Optional[bool] = True
+    is_superuser: Optional[bool] = False
+    full_name: Optional[str] = None
+
+class UserBaseInDB(UserBase):
+    id: int = None
+
+
+# Properties to receive via API on creation
+class UserCreate(UserBaseInDB):
+    email: str
+    password: str
+
+
+# Properties to receive via API on update
+class UserUpdate(UserBaseInDB):
+    password: Optional[str] = None
+
+
+# Additional properties to return via API
+class User(UserBaseInDB):
+    pass
+
+
+# Additional properties stored in DB
+class UserInDB(UserBaseInDB):
+    hashed_password: str
